Enhance registration UX and stats, update assets and styles
Refactored restaurant registration flow with a new confirmation step, improved opening hours validation, and enhanced geolocation UX. Updated PlatStatsOverview to show real-time stats, fixed Commande_plat table name, and changed restaurant latitude/longitude to double. Improved various frontend styles, removed unused favicon assets, and updated copyright.

+++ apps/admin/database/migrations/2025_08_28_112407_create_restaurants_table.php
@@ -22,8 +22,8 @@ return new class extends Migration
             $table->string('name');
             $table->string('slug')->unique();
             $table->text('description')->nullable();
-            $table->decimal('latitude', 10, 7)->nullable();
-            $table->decimal('longitude', 10, 7)->nullable();
+            $table->double('latitude')->nullable();
+            $table->double('longitude')->nullable();
             $table->string('logo_path')->nullable();
             $table->string('phone')->nullable();
             $table->string('ville')->nullable();
